requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
django-filter==25.1
drf-spectacular==0.28.0
psycopg2-binary==2.9.10
celery==5.5.3
redis==5.0.1
Pillow==11.3.0

# Additional packages for production
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0

# Development packages
django-debug-toolbar==4.4.6
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1

# SMS and Email
twilio==9.3.7
boto3==1.35.91

# DigitalOcean Spaces / S3 Storage
django-storages==1.14.6

# Payment gateways
stripe==12.2.0
razorpay==1.4.2

# Utilities
python-decouple==3.8
phonenumbers==8.13.52
qrcode==8.0
reportlab==4.2.5
requests==2.31.0

channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

